04-embedder/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Pydantic for data validation
pydantic==2.5.0

# Google Generative AI
google-genai

# Message queue
pika==1.3.2

# Utilities
python-dotenv==1.0.0